requirements-dev.txt

black
flake8
flask-shell-ipython
ipdb
ipython
isort
pytest
pytest-flask
pytest-cov
python-dotenv
Flask-Migrate